+++ src/extensions/session/sessionext.cpp
@@ -46,7 +46,9 @@
  */
 SessionClassExtension::SessionClassExtension(const SessionClass *this_ptr) :
     GlobalExtensionClass(this_ptr),
-    ExtOptions()
+    ExtOptions(),
+    IsChatToAllies(false),
+    MessageRecipientName("")
 {
     //if (this_ptr) EXT_DEBUG_TRACE("SessionClassExtension::SessionClassExtension - 0x%08X\n", (uintptr_t)(ThisPtr));