+++ src/new/ui/uicontrol.h
@@ -32,6 +32,8 @@
 #include "point.h"
 #include "typelist.h"
 
+#include <string>
+
 
 struct IStream;
 class CCINIClass;
@@ -309,6 +311,28 @@ class UIControlsClass
          *  Should the sidebar repair, etc. buttons use the old X position, centered on the radar?
          */
         bool IsCenterSidebarButtonsOnRadar;
+
+        /**
+         *  Beacon animations are not tied to game FPS, this is the FPS at which they play.
+         */
+        int BeaconAnimFramesPerSecond;
+        int RadarBeaconAnimFramesPerSecond;
+
+        /**
+         *  The offset at which beacon text is drawn (on actual beacons, and during preview).
+         */
+        int BeaconTextOffset;
+        int BeaconPreviewTextOffset;
+
+        /**
+         *  Text presets for beacons when holding modifier keys when placing.
+         */
+        std::string BeaconText[7];
+
+        /**
+         *  Text presets for beacons *shown as a tooltip* when holding modifier keys when placing.
+         */
+        std::string BeaconPreviewText[7];
 };
 
 extern UIControlsClass *UIControls;